prevent other users from deleting a blog that is created by other users

+++ controllers/blogs.js
@@ -45,6 +45,15 @@ blogsRouter.post('/', async (request, response) => {
 })
 
 blogsRouter.delete('/:id', async (request, response) => {
+  const decodedToken = jwt.verify(request.token, process.env.SECRET)
+  if (!decodedToken?.id) {
+    return response.status(401).json({ error: 'token invalid' })
+  }
+
+  const blogToDelete = await Blog.findById(request.params.id)
+  if (decodedToken.id.toString() !== blogToDelete.user.toString()) {
+    return response.status(401).json({ error: 'you are not authorized to remove this blog' })
+  }
   await Blog.findByIdAndRemove(request.params.id)
   response.status(204).end()
 })
